+++ rust/src/tests/mocks/rpc/mod.rs
@@ -1,0 +1,1 @@
+pub mod aggregator;
