+++ rust/src/tests/mod.rs
@@ -1,1 +1,2 @@
 mod coordinator;
+pub mod mocks;
